Finish daily tax implementation
Everything should be ret-2-go

+++ common/src/commonMain/kotlin/net/perfectdreams/loritta/cinnamon/common/utils/DailyTaxThresholds.kt
@@ -6,29 +6,29 @@ object DailyTaxThresholds {
      */
     val THRESHOLDS = listOf(
         DailyTaxThreshold(
-            3L,
+            3,
             100_000_000L,
             0.5
         ),
         DailyTaxThreshold(
-            7L,
+            7,
             10_000_000L,
             0.25
         ),
         DailyTaxThreshold(
-            14L,
+            14,
             1_000_000L,
             0.1
         ),
         DailyTaxThreshold(
-            30L,
+            30,
             100_000L,
             0.05
         )
     ).sortedByDescending { it.minimumSonhosForTrigger }
 
     data class DailyTaxThreshold(
-        val maxDayThreshold: Long,
+        val maxDayThreshold: Int,
         val minimumSonhosForTrigger: Long,
         val tax: Double
     )
